+++ NEWS.md
@@ -1,3 +1,7 @@
+- Internal
+  - Replace parameters() with pars()
+  - Replace subset_mcmcr_parameters() with subset()
+
 # mcmcderive 0.0.1
 
 * Optimized `mcmc_derive()`.